refactor: decouple Modal component from App, pass addToCart handler instead of setState prop
